requirements.txt
jupyter-book
autodoc_pydantic
./bsmschema
